+++ Faker.Tests/BusinessTests.cs
@@ -31,5 +31,17 @@ namespace RimuTec.Faker.Tests {
          Assert.AreEqual(3, RegexMatchesCount(creditCardNumber, @"-"));
          Assert.AreEqual(4, RegexMatchesCount(creditCardNumber, @"[0-9]{4}"));
       }
+
+      [Test]
+      public void CreditCardType_HappyDays() {
+         // arrange
+
+         // act
+         var type = Business.CreditCardType();
+
+         // assert
+         Assert.IsTrue("['visa', 'mastercard', 'american_express', 'discover', 'diners_club', 'jcb', 'switch', 'solo', 'dankort', 'maestro', 'forbrugsforeningen', 'laser']"
+            .Contains($"'{type}'"));
+      }
    }
 }